requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.4.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
comm==0.1.2
cryptography==36.0.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.0
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.11
django-cors-headers==3.10.1
django-filter==22.1
django-heroku==0.3.1
django-rest-knox==4.1.0
django-upgrade==1.4.0
djangorestframework==3.12.4
dynaconf==3.1.7
executing==1.2.0
filelock==3.4.0
firebase-admin==6.1.0
Flask==2.2.2
fonttools==4.29.1
frozenlist==1.3.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
h11==0.13.0
httplib2==0.21.0
hvac==0.11.2
idna==3.3
ipykernel==6.21.3
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jwcrypto==1.4.2
kiwisolver==1.3.2
knox==0.1.14
kombu==5.2.4
lightgbm==3.3.2
loguru==0.5.3
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
msgpack==1.0.4
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.22.2
oauth2client==4.1.3
openai==0.27.2
outcome==1.1.0
packaging==21.3
pandas==1.4.4
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.31
proto-plus==1.22.2
protobuf==4.22.0
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
Pyrebase4==4.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2021.3
pyzmq==25.0.1
requests==2.26.0
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
selenium==4.1.2
shortuuid==1.0.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.6.2
threadpoolctl==3.1.0
tokenize-rt==4.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==5.3.0
win32-setctime==1.0.4
wrapt==1.14.1
wsproto==1.1.0
yarl==1.8.2
